+++ src/app/SBE/detail-profile/detail-layout/helpers/get-detail-type.js
