+++ src/app/SBE/detail-profile/detail-layout/helpers/set-top-height.js
