+++ src/app/SBE/detail-profile/detail-layout/intern/detail-sections/desktop/RightDetailSection.js
@@ -1,12 +1,12 @@
 import { VStack } from '@chakra-ui/layout'
 
 import 'app/layouts/components/css/hide-scroll.css'
-import Header from 'app/SBE/detail-profile/template/layouts/Header.js'
+import Header from 'app/SBE/detail-profile/detail-layout/template/layouts/Header.js'
 import {
   InternshipPreferenceSection,
   InternshipSection,
   ExperienceSection,
-} from 'app/SBE/detail-profile/template/layouts/ProfileSections.js'
+} from 'app/SBE/detail-profile/detail-layout/template/layouts/ProfileSections.js'
 
 const RightDetail = ({ beCode }) => {
   return (
