+++ src/app/SBE/detail-profile/detail-layout/template/attributes-list/index.js
