+++ src/app/SBE/detail-profile/detail-layout/template/layouts/Actions.js
