+++ src/app/SBE/detail-profile/detail-layout/template/layouts/ProfileSections.js
@@ -10,10 +10,10 @@ import {
   prefs,
   internshipDetails,
   recentEmployment,
-} from 'app/SBE/detail-profile/template/attributes-list/index.js'
+} from 'app/SBE/detail-profile/detail-layout/template/attributes-list/index.js'
 
 export const InternshipPreferenceSection = ({ beCode }) => {
-  const cardBg = useColorModeValue('blue.200', 'gray.600')
+  const cardBg = useColorModeValue('gray.200', 'gray.600')
   return (
     <Center w="full">
       <Card p={6} w="full" bg={cardBg} overflow="hidden" borderRadius="2rem">
